Disable handle_commands
Disable handle_commands so that the constant "!xyz is not a code" error stops showing. This was not causing any problems, but blogged up the console with errors that don't mean anything. This happened because the bot.handle_commands(ctx) was trying to find a corresponding command to messages starting with ! symbol.